Rename ps2_psh.c to vsv.c
